feat: shifted fiels to proper directories
