+++ src/app.py
@@ -2,7 +2,7 @@ from typing import Annotated
 from fastapi import FastAPI, Depends
 from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
 from utils.verifyaccesstoken import VerifyAccessTokenUtil, VerifiedToken, JwtSubjectType
-from routes import health, pdf, video, stream_chat, axa_travel_entitlements_chat, axa_travel_policy_chat, finance_report_gen
+from routes import health, pdf, video, stream_chat, axa_travel_entitlements_chat, finance_report_gen
 from asgi_correlation_id import CorrelationIdMiddleware
 from fastapi.middleware.cors import CORSMiddleware
 
@@ -38,7 +38,7 @@ app.include_router(pdf.router, prefix='/api', dependencies=[Depends(validateToke
 app.include_router(video.router, prefix='/api', dependencies=[Depends(validateToken)], tags=["Video"])
 
 app.include_router(stream_chat.router, prefix='/api', dependencies=[Depends(validateToken)], tags=["StreamChat"])
+
 app.include_router(axa_travel_entitlements_chat.router, prefix='/api', dependencies=[Depends(validateToken)], tags=["Axa business travel entitlements Chat"])
-app.include_router(axa_travel_policy_chat.router, prefix='/api', dependencies=[Depends(validateToken)], tags=["Axa business travel policy Chat"])
 
 app.include_router(finance_report_gen.router, prefix='/api', dependencies=[Depends(validateToken)], tags=["Annual Report Financial Report Generation"])
