+++ src/configs/__init__.py
@@ -1,2 +1,3 @@
 from .configs import *
 from .latex import *
+from .prompts import *
