+++ src/configs/prompts/__init__.py
@@ -1,0 +1,2 @@
+from .annual_report import *
+from .axa_travel_chat import *
