+++ src/configs/prompts/annual_report/annual_report_products_section.py
